Update admin impersonation specs
I removed the ChA impersonation specs, these specs were commented out
back in 2017: 35efaf398fa0090099bcc43f5cbc40cee16bbfd3

I'm guessing ChAs used to be able to impersonate, but then at some point
that functionality got removed. Anywho, the specs aren't necessary and
we can always add them back if this functionality comes back.

I also did a little clean up and renamed this test file.
